biet_web/requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.8
django-phonenumber-field==4.0.0
gunicorn==20.0.4
phonenumbers==8.12.6
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
uWSGI==2.0.19.1
